Created two packages: commom and video-processor
The commom package will have commom code that can be used in other packages

the main file in commom package has auxiliary functions for functional programming.

bind and asyncBind functions can be used to transform a function that receives a value into a function which is
prepared to receive a value or an error and process the value or pass the error.

map and asyncMap is very similar to bind, but the function transformed was not expecting an error.

tee and asyncTee is used to execute a void function and continue with the workflow.

In commom, It was defined some types for commom services, such as SaveToCDN, SaveToDatabase and SaveToStorage.

It was implemented the CheckFileExists service, which is used to check if a file exists in a storage.

In VideoProcessor package, it was implemented the UploadVideo Workflow, which will receive information about a file
in a request HTTP and will upload to a storage.
It also was implemented the GenerateUnprocessedVideo Workflow, which will receive the information about a already saved to
storage file and will generate a unprocessed video and save it to database.
It was implemented the ProcessVideo Workflow, which will receive the information about a unprocessed video and will generate
the different resolutions.

It will be necessary to implement the service which will generate the different resolutions.

In this workflow, It will be necessary to delete information about the unprocessed video from database.

Tests were not implemented yet.